+++ .changes/stream-package.md
@@ -1,0 +1,11 @@
+---
+"@effection/subscription": "minor"
+"@effection/atom": "minor"
+"@effection/channel": "minor"
+"@effection/duplex-channel": "minor"
+"@effection/events": "minor"
+"@effection/process": "minor"
+"effection": "minor"
+---
+
+Split off `Stream` from subscription package into its own `@effection/stream` package
