+++ packages/duplex-channel/package.json
@@ -40,7 +40,7 @@
   "dependencies": {
     "@effection/core": "^2.0.0-beta.10",
     "@effection/channel": "^2.0.0-beta.12",
-    "@effection/subscription": "^2.0.0-beta.12"
+    "@effection/stream": "^2.0.0-beta.1"
   },
   "volta": {
     "extends": "../../package.json"
