+++ packages/effection/package.json
@@ -33,7 +33,8 @@
     "@effection/events": "2.0.0-beta.16",
     "@effection/fetch": "2.0.0-beta.17",
     "@effection/main": "2.0.0-beta.16",
-    "@effection/subscription": "2.0.0-beta.16"
+    "@effection/subscription": "2.0.0-beta.16",
+    "@effection/stream": "2.0.0-beta.1"
   },
   "devDependencies": {
     "@babel/core": "^7.14.8",
