+++ packages/events/package.json
@@ -29,10 +29,11 @@
   },
   "dependencies": {
     "@effection/core": "2.0.0-beta.14",
-    "@effection/subscription": "2.0.0-beta.16"
+    "@effection/stream": "2.0.0-beta.1"
   },
   "devDependencies": {
     "@effection/mocha": "2.0.0-beta.9",
+    "@effection/subscription": "2.0.0-beta.16",
     "@frontside/tsconfig": "^1.2.0",
     "@types/node": "16.3.2",
     "expect": "^25.4.0",
