+++ packages/stream/README.md
@@ -1,0 +1,4 @@
+# @effection/stream
+
+APIs for producing, consuming and transforming streams of data within
+effection operations.
