+++ packages/stream/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@effection/process",
-  "version": "2.0.0-beta.16",
-  "description": "Spawn and manage external processes with Effection",
+  "name": "@effection/stream",
+  "version": "2.0.0-beta.1",
+  "description": "Effection Stream",
   "main": "dist-cjs/index.js",
   "module": "dist-esm/index.js",
   "types": "dist-esm/index.d.ts",
@@ -11,7 +11,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/thefrontside/effection.git",
-    "directory": "packages/process"
+    "directory": "packages/stream"
   },
   "author": "Frontside Engineering <engineering@frontside.com>",
   "license": "MIT",
@@ -23,30 +23,22 @@
   ],
   "scripts": {
     "lint": "eslint '{src,tests}/**/*.ts'",
-    "test": "mocha -r ts-node/register --timeout 10000 test/**/*.test.ts",
+    "test": "mocha -r ts-node/register --timeout 5000 test/**/*.test.ts",
     "prepack": "tsc --build tsconfig.esm.json && tsc --build tsconfig.cjs.json",
     "mocha": "mocha -r ts-node/register"
+  },
+  "dependencies": {
+    "@effection/core": "2.0.0-beta.14",
+    "@effection/subscription": "2.0.0-beta.16"
   },
   "devDependencies": {
     "@effection/mocha": "2.0.0-beta.9",
     "@frontside/tsconfig": "^1.2.0",
-    "@types/cross-spawn": "^6.0.2",
-    "@types/node": "^16.3.2",
-    "@types/node-fetch": "^2.5.7",
+    "@types/mocha": "^8.0.3",
     "expect": "^25.4.0",
     "mocha": "^8.3.1",
-    "node-fetch": "^2.6.1",
     "ts-node": "^10.1.0",
     "typescript": "^4.3.5"
-  },
-  "dependencies": {
-    "@effection/channel": "2.0.0-beta.16",
-    "@effection/core": "2.0.0-beta.14",
-    "@effection/events": "2.0.0-beta.16",
-    "@effection/stream": "2.0.0-beta.1",
-    "cross-spawn": "^7.0.3",
-    "ctrlc-windows": "^2.0.0",
-    "shellwords": "^0.1.1"
   },
   "volta": {
     "extends": "../../package.json"
