+++ packages/stream/src/match.ts
