+++ packages/stream/src/symbol-operation-iterable.ts
