+++ packages/stream/src/writable-stream.ts
