+++ packages/stream/tsconfig.cjs.json
@@ -1,0 +1,10 @@
+{
+  "extends": "./tsconfig",
+  "compilerOptions": {
+    "rootDir": "./src",
+    "outDir": "./dist-cjs",
+    "noEmit": false
+  },
+  "references": [{ "path": "../core/tsconfig.cjs.json" }],
+  "exclude": ["./test/*"]
+}
