+++ packages/stream/tsconfig.esm.json
@@ -1,0 +1,9 @@
+{
+  "extends": "./tsconfig.cjs.json",
+  "compilerOptions": {
+    "outDir": "./dist-esm",
+    "module": "ESNext"
+  },
+  "references": [{ "path": "../core/tsconfig.esm.json" }],
+  "exclude": ["./test/*"]
+}
