+++ packages/subscription/src/index.ts
@@ -1,7 +1,3 @@
-export { SymbolOperationIterable } from './symbol-operation-iterable';
-export { OperationIterable } from './operation-iterable';
 export { OperationIterator } from './operation-iterator';
 export { Subscription } from './subscription';
-export { createStream, Stream, StringBufferStream } from './stream';
-export { Writable, WritableStream } from './writable-stream';
 export { Queue, createQueue } from './queue';
